pyproject.toml
[project]
name = "marketingteam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-openai>=1.1.3",
    "langchain-anthropic>=1.3.0",
    "python-dotenv>=1.2.1",
]
